ComponentAgent correctly updates Speed and Acceleration
In order to avoid the "bounce back" effect, a high acceleration must be set to the agent (p.eg: 999). Max speed will be the regular speed.

Take into account that changing acceleration may not fix the bounce effect if speed is too high. If it's the case, acceleration will have to be set to a higher ammount. This will be fixed in future evolutionaries.